Extract temporary-defined types to its own file
